Adicionando o trabalho 2 e algumas estruturas de dados feitas

+++ extras/testeLong.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    printf("%ld\n", sizeof(long double));
+    printf("%ld\n", sizeof(long));
 
     return 0;
 }
